+++ changelog.rst
@@ -6,6 +6,7 @@ Features:
 
 * Toggle autocommit mode using F6 or a config setting.
 * When in a transaction, rollback erring queries without aborting the transaction. ('On error rollback'.)
+* Less verbose bottom bar.
 
 Bug fixes:
 ----------